third_party/blink/web_tests/http/tests/websocket/url-with-credential-expected.txt

Make sure handshake with credentials in URL is successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS host is "127.0.0.1:8880"
PASS successfullyParsed is true

TEST COMPLETE
